test-wayland-surface: Adds basic xwayland support
This adds basic support for spawning a headless xwayland server when
running test-wayland-surface so that legacy X clients can be composited.

Since this depends on the xserver extension protocol which currently
lives in the wayland-demos repository this patch adds a
--with-wayland-protocols ./configure option so clutter can be told where
to find the xml for the xserver extension xml.

FIXME: This patch is currently hard coding the patch for xwayland to be
/home/bob/local/xserver-xwayland/bin/X