test-wayland-surface: Update to the new shell interface
Wayland has changed so that the shell interface now only has one
function which returns a shell surface for the surface. This patch makes
it create a dummy service in the same way that the wayland demo
compositor does. The implementation of the shell_surface_interface for
that dummy service is all no-ops.  ate test-wayland-surface to support
latest protocol

This patch was based on an equivalent patch made to cogland by Neil
Roberts.