cogland: Fix crash when frame callback is destroyed before emitted
If a frame callback is destroyed before it is invoked then the struct
would be freed but it would not be removed from the array of callbacks
so when cogland later tried to emit the callback it would crash. This
patch instead stores the callbacks in a GQueue with embedded list nodes
so that they can be removed from the list in the resource destructor.
That way it doesn't matter how the resource is destroyed, it will still
get removed from the list.

This patch was based on an equivalent patch made to cogland by Neil
Roberts.