Fix the logic for opening the non-abstract socket (again)
The change in commit be1bc97ca437 misunderstood the intention of
opening the non-abstract socket by assuming it was a fallback. In fact
the code should be opening both types of socket. It later passes both
of them down to XWayland by an event in the Wayland protocol. This was
causing it to pass down a bogus file descriptor for the non-abstract
socket which somehow made it call accept on the TTY device which would
make the X server get stuck in an infinite loop if any input came in
on the console.